# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Royal Dental Centre API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: doctorId
          in: query
          required: true
          schema:
            type: integer
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (doctor/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Announcement deleted
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
      required:
        - firstName
        - lastName
        - email
        - password
        - phone
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [patient, doctor, admin]
        dateOfBirth:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - role
        - createdAt
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        serviceId:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        notes:
          type: ["string", "null"]
        patientName:
          type: string
        doctorName:
          type: string
        serviceName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - doctorId
        - serviceId
        - date
        - time
        - status
        - patientName
        - doctorName
        - serviceName
        - createdAt
    CreateAppointmentBody:
      type: object
      properties:
        doctorId:
          type: integer
        serviceId:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
        notes:
          type: string
      required:
        - doctorId
        - serviceId
        - date
        - time
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        notes:
          type: string
        date:
          type: string
        time:
          type: string
    TimeSlot:
      type: object
      properties:
        time:
          type: string
        available:
          type: boolean
      required:
        - time
        - available
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        type:
          type: string
          enum: [news, offer, update]
        isActive:
          type: boolean
        authorName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - titleAr
        - content
        - contentAr
        - type
        - isActive
        - authorName
        - createdAt
    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        content:
          type: string
        contentAr:
          type: string
        type:
          type: string
          enum: [news, offer, update]
      required:
        - title
        - titleAr
        - content
        - contentAr
        - type
...