# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Royal Dental Centre API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  /** @nullable */
  dateOfBirth?: string | null;
  createdAt: Date;
}